Make clang-tidy happy

+++ src/fheroes2/agg/agg_image.cpp
@@ -4015,7 +4015,7 @@ namespace fheroes2
                 if ( !_icnVsSprite[id].empty() ) {
                     Sprite & sprite = _icnVsSprite[id][0];
                     uint8_t * image = sprite.image();
-                    const uint8_t * imageEnd = image + sprite.width() * sprite.height();
+                    const uint8_t * imageEnd = image + static_cast<size_t>( sprite.width() ) * sprite.height();
                     const uint8_t * transform = sprite.transform();
 
                     for ( ; image != imageEnd; ++image, ++transform ) {
